move Maven SettingsProvider to core
