+++ core/src/main/java/hudson/mvn/FilePathSettingsProvider.java
@@ -1,20 +1,21 @@
-package hudson.maven;
+package hudson.mvn;
 
 import hudson.Extension;
 import hudson.FilePath;
 import hudson.model.AbstractBuild;
+import hudson.model.AbstractProject;
 import hudson.util.ArgumentListBuilder;
 import hudson.util.FormValidation;
 import hudson.util.IOUtils;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+
 import org.apache.commons.lang.StringUtils;
 import org.kohsuke.stapler.AncestorInPath;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.QueryParameter;
-
-import javax.servlet.ServletException;
-import java.io.IOException;
-
-import static hudson.Util.fixEmpty;
 
 /**
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
@@ -54,7 +55,7 @@ public class FilePathSettingsProvider extends SettingsProvider {
         /**
          * Check that the provided file is a relative path. And check that it exists, just in case.
          */
-        public FormValidation doCheck(@AncestorInPath MavenModuleSet job, @QueryParameter String value) throws IOException, ServletException {
+        public FormValidation doCheck(@AncestorInPath AbstractProject job, @QueryParameter String value) throws IOException, ServletException {
             String v = fixEmpty(value);
             if ((v == null) || (v.length() == 0)) {
                 // Null values are allowed.
@@ -64,7 +65,7 @@ public class FilePathSettingsProvider extends SettingsProvider {
                 return FormValidation.error(Messages.MavenModuleSet_AlternateSettingsRelativePath());
             }
 
-            MavenModuleSetBuild lb = job.getLastBuild();
+            Run lb = job.getLastBuild();
             if (lb!=null) {
                 FilePath ws = lb.getWorkspace();
                 if(ws!=null)
