+++ core/src/main/resources/hudson/mvn/DefaultSettingsProvider/config.jelly
