+++ core/src/main/resources/hudson/mvn/DefaultSettingsProvider/help.html
