+++ core/src/main/resources/hudson/mvn/FilePathSettingsProvider/help-path.html
