+++ core/src/main/resources/hudson/mvn/FilePathSettingsProvider/help.html
