+++ maven-plugin/src/main/java/hudson/maven/MavenModuleSet.java
@@ -56,6 +56,10 @@ import hudson.model.SCMedItem;
 import hudson.model.Saveable;
 import hudson.model.TaskListener;
 import hudson.model.TopLevelItem;
+import hudson.mvn.DefaultSettingsProvider;
+import hudson.mvn.FilePathSettingsProvider;
+import hudson.mvn.SettingsProvider;
+import hudson.mvn.SettingsProviderDescriptor;
 import hudson.search.CollectionSearchIndex;
 import hudson.search.SearchIndexBuilder;
 import hudson.tasks.BuildStep;
