+++ maven-plugin/src/main/java/hudson/maven/MavenModuleSetBuild.java
@@ -53,6 +53,7 @@ import hudson.model.Result;
 import hudson.model.Run;
 import hudson.model.StringParameterDefinition;
 import hudson.model.TaskListener;
+import hudson.mvn.SettingsProvider;
 import hudson.remoting.VirtualChannel;
 import hudson.scm.ChangeLogSet;
 import hudson.tasks.BuildStep;
